+++ execute/executeOptions.go
@@ -6,7 +6,7 @@ import (
 	"github.com/apenella/go-ansible/stdoutcallback"
 )
 
-// Executor interface is satisfied by those types which has a Execute(context.Context,string,[]string)error method
+// Executor interface is satisfied by those types which has a Execute(context.Context,[]string,stdoutcallback.StdoutCallbackResultsFunc,...ExecuteOptions)error method
 type Executor interface {
 	Execute(ctx context.Context, command []string, resultsFunc stdoutcallback.StdoutCallbackResultsFunc, options ...ExecuteOptions) error
 }