+++ stdoutcallback/stdoutCallback.go
@@ -8,8 +8,8 @@ import (
 	"github.com/apenella/go-ansible/stdoutcallback/results"
 )
 
-// StdoutCallbackResultsFunc defines a function which manages ansible's stdout callbacks. The function expects and string for prefixing output lines, a reader that receives the data to be wrote and a writer that defines where to write the data comming from reader
-type StdoutCallbackResultsFunc func(context.Context, string, io.Reader, io.Writer) error
+// StdoutCallbackResultsFunc defines a function which manages ansible's stdout callbacks. The function expects a context, a reader that receives the data to be wrote and a writer that defines where to write the data comming from reader, Finally a list of transformers could be passed to update the output comming from the executor.
+type StdoutCallbackResultsFunc func(context.Context, io.Reader, io.Writer, ...results.TransformerFunc) error
 
 const (
 	// AnsibleStdoutCallbackEnv